ai proxy fix
