+++ apps/backend/src/app/api/latest/integrations/ai-proxy/[[...path]]/route.ts
@@ -1,4 +1,6 @@
 import { ALLOWED_MODEL_IDS } from "@/lib/ai/models";
+import { collectAllowedProxyReferences } from "@/lib/ai/proxy-preprocessing";
+import { preprocessProxyBody } from "@/private";
 import { handleApiRequest } from "@/route-handlers/smart-route-handler";
 import { getEnvVariable } from "@stackframe/stack-shared/dist/utils/env";
 import { StatusError } from "@stackframe/stack-shared/dist/utils/errors";
@@ -29,6 +31,11 @@ function sanitizeBody(raw: ArrayBuffer): Uint8Array {
   if (parsed.metadata?.user_id && parsed.metadata.user_id.length > 128) {
     parsed.metadata.user_id = parsed.metadata.user_id.slice(0, 128);
   }
+
+  parsed = preprocessProxyBody({
+    parsedBody: parsed,
+    allowedReferences: collectAllowedProxyReferences(),
+  });
 
   return new TextEncoder().encode(JSON.stringify(parsed));
 }
