+++ apps/backend/src/private/implementation-fallback/index.ts
@@ -1,3 +1,4 @@
+import { AiProxyBodyProcessor } from "@/lib/ai/proxy-preprocessing";
 import { SignUpRiskEngine } from "@/lib/risk-scores";
 import { createNeutralSignUpHeuristicFacts } from "@/lib/sign-up-heuristics";
 
@@ -9,3 +10,5 @@ export const signUpRiskEngine: SignUpRiskEngine = {
     };
   },
 };
+
+export const preprocessProxyBody: AiProxyBodyProcessor = ({ parsedBody }) => parsedBody;
